+++ src/possible-routes.js
@@ -156,8 +156,8 @@ function iterateRoutes(endList, graphEdgeArr, critListArr, ignoreCrit, routeArra
 		{
 			// Derive new routes from possible destinations.
 			currentAdjEdges = readIncompleteRoute(currentNode, currentRoute.distance, currentStops, critListArr, ignoreCrit, graphEdgeArr);
-			routeTasks.deriveNew(routeIndex, currentRoute, currentAdjEdges, graphEdgeArr, routeArray, allowBacktracking);
-			currentOffset = currentAdjEdges.length;
+			currentOffset = routeTasks.deriveNew(routeIndex, currentRoute, currentAdjEdges, graphEdgeArr, routeArray, allowBacktracking);
+			//currentOffset = currentAdjEdges.length;
 		}